Update travelGraph branch

+++ src/travelGraph/TravelGraph.java
@@ -33,7 +33,7 @@ public class TravelGraph {
 	 * @param duration The duration of the trip in hours
 	 * @return The updated path. Returns a new path if no path existed with the given origin, destination, company name, transport type.
 	 */
-	public Path updatePathPrice(String origin, String destination, String company, Path.TransportType type, float ppg, float ppcc, float maxWeight, float maxVol, Path.DayOfWeek dayOfWeek, int frequency, int duration){
+	public Path updatePathPrice(String origin, String destination, String company, Path.TransportType type, float ppg, float ppcc, int maxWeight, int maxVol, Path.DayOfWeek dayOfWeek, int frequency, int duration){
 		Location[] locs = getLocs(origin, destination);
 		Location originLoc = locs[0];
 		Location destLoc = locs[1];
